Adds method to get photo urls
This commit adds a function that, given a rover name, date, and API Key,
retrieves a list of photo urls for photos taken by the rover on the
specified date.

The function has very limited error checking. The program is hard-coded
to always retrieve photos for the Mars Curiosity rover. But this is just
a sample app to test working with the API.