Refactors argument processing; handles date file
This commit ended up doing two things...probably should be two separate
commits.

First, argument processing was moved out of the Main method. Each
argument was given a "helper extension" method in the Helpers class that
encapsulates the processing of that particular argument.

One notable thing is that --date and --dates is processed by the same
helper method because the arguments are mutually exclusive and both end
up populating a single List<DateTime> which is used to retrieve the
photos for a given rover for the given dates. Again, the rover name is
hard-coded to "curiosity". Curiously (pun intended), there are no photos
for Curiousty rover on 13-Jul-2016. The next commit will address this.

And as alluded to, the program can now process a list of dates given a
text file containing a list of line-separated dates. Dates which cannot
be parsed are simply ignored.

Finally, Program.cs was getting to big; the Helpers class has gained a
lot of helper methods. So they were moved to their own file. And given
all the methods created to work with the program's arguments, a unit
test project has been created to enuser propery argument processing. Lo
and behold, there were a few things the tests found. That's why you
test.