Downloads all photos for all rovers on specified dates
This commit allows a sequence of rovers to be passed to the
DownloadPhotos method (actually, it's a new overload).

For each given date, the api is queried for each rover (the rover is
part of the path segment in the API) and if any results are returned by
the API for a given rover on a given date, those images are retrieved.

With this, the full "demo" app is complete and it can be "deconstructed"
into a Web API for downloading and caching NASA Mars Rover photos.

I'm considering the following functionality:

  - Given a pre-set lit of dates, upon application startup,
    download all tho photos for the given dates to a configurable
    location, if they haven't already been downloaded.

    - Provide an endpoint to allow specifying a rover name and a date
      returning a list of photo filenames and their URL.

      - If there are no cached photos on the specified date for the
        specified rover:

        - If the file NoPhotos.txt exists, send back a 204 No Content
          e.g. the request was successful, but there are no photos.

        - Otherwise, attempt to retrieve the photos and cache them locally
          - If no photos exist on the given date, create a file:
            NoPhotos.txt