Adds application configuration
This commit adds configuration to appsettings.json for configuring the
WebAPI. In addition, I implemented loading that configuration using a
set of configuration classes and extension methods that is inline with
the ASP.NET Core configuration ecosystem/API.

NOTE: In order to demonstrate that this configuration was working,
controller's '/' action method WILL display the configured API key (if
set--and if not set in User Secrets, it'll be blank at this time). A
future commit will remove this code.