Adds the actual functionality to action methods
This commit implements the RoverPhotosController action methods. I was
planning on several commits, but it turns out that all of the action
methods really used very similar queries.

When the Web API starts, it reads the dates.txt file and retrieves the
list of all photos that are available for each rover on a specified
dates. (Malformed dates are ignored.)

The photo downloader will not re-download a file, So the initialization
hit is only experienced once, on first application startup. There is a
little hit to query the NASA API for the list of photos on each date on
each application startup--but this ensures that if a new date is added,
new photos will be downloaded (if any exist).