WIP: React app hits WebAPI to load carousel on home page
This is a WIP commit. The home page content is mostly complete and loads 4
random photos into a carousel.

Next steps:

* Add "cards" to the home page, one for each rover with photos.
    * Grab a random photo from each rover to display in the card

* Finish "search" functionality.
    * Requires creation of a Search component/page

* Create pages for each rover which displays some number of cards with a
photo.
    * Integrate lightbox to display a magnified view of the photo?
    * Implement "inifinite scrolling" for the photos?
    * These ideas could be used on the search page.
    * Come to think of it, there isn't much difference between the
    search pages and the individual rover pages other than their title
    and the content retrieved, so this is a good candidate for a generic
    component.

Items to improve:

* Probably need to modify the WebAPI to give me a total count of
available photos. Right now, I used Windows Explorer to find this number
and hard-coded the value which is used to determine the upper-bound of a
random number to be generated that indicates how many "pages" of results
should be skipped.

* Even better than taking 'x' consecutive photos, take 'x' random
photos. I noticed that often times, a larger photo is selected with an
identical smaller photo. Not very interesting.

* Setup Docker or Kubernetes so all you need to do is 'docker-compose
up' or the equivalent k8s command.

To run, start the Web API in Visual Studio. Then 'yarn start' in this
project's directory. Then open your browser to the URL shown in the
console.