Adds a card deck to the home page
This commit adds a "card deck" to the home page that showcases a random
photo from a given rover. Each card has a button prompting the user to
click to view more photos for that rover. The entire card is one giant
link.

As part of creating this UI, a function which extracts the rover name
from the image URL was moved out of the carousel photo component back to
the HomePage component, and subsequently passed to the carousel
component--so that I could also use the function on the home page.

All that's really left in terms of UI work is to implement the "results"
component that will display a page for specific rover photos or, if a
search was performed, the search results.