WIP: Adds rover photos pages w/lightbox
This commit adds a PhotosPage component that displays a grid of cards
for each photo for a rover (or for the specified search
criteria--though, searching is not working yet).

Currently, the pages do not paginate their results. All photos for a
specified rover are displayed. You can click on a rover photo to open a
lightbox which allows you to zoom in on the photo and navigate to the
previous/next photo.

If no photos exist for a rover, the page displays a No Photos message.

Known Issues: The spinner is not centered in the page.